+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.21.0] - 2024-02-08
+
 ### Added
 
 - Add `ContextSymbol.allowedToReferenceSymbols`.
@@ -1388,6 +1390,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version 1.0.0
 
 [unreleased]: https://github.com/Decompollaborate/spimdisasm/compare/master...develop
+[1.21.0]: https://github.com/Decompollaborate/spimdisasm/compare/1.20.1...1.21.0
 [1.20.1]: https://github.com/Decompollaborate/spimdisasm/compare/1.20.0...1.20.1
 [1.20.0]: https://github.com/Decompollaborate/spimdisasm/compare/1.19.0...1.20.0
 [1.19.0]: https://github.com/Decompollaborate/spimdisasm/compare/1.18.0...1.19.0